refactor: move Output base class into outputs/base.py; thin outputs/__init__
outputs/__init__.py held the entire ~140-line Output base class. Extract it to
outputs/base.py and reduce __init__ to a facade (re-export Output, LSLOutlet,
UDPOutput). Concrete outputs now import Output from outputs.base (not the
package) to avoid an init-order cycle. Public API unchanged.
