+++ myogestic/outputs/serial_output.py
@@ -4,7 +4,7 @@ import struct
 
 import numpy as np
 
-from myogestic.outputs import Output
+from myogestic.outputs.base import Output
 
 
 class SerialOutput(Output):
